requirements.txt
fastapi>=0.110.3
inference_gpu==0.23.0
matplotlib==3.8.1
numpy>=1.26.1
opencv_contrib_python==4.9.0.80
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
passlib==1.7.4
pox==0.3.5
pydantic==2.9.2
python_firebase==1.2.1
python_jose==3.3.0
Requests==2.32.3
tensorflow==2.16.1
tensorflow==2.14.0
tensorflow_hub==0.16.1
tensorflow_intel==2.16.1
tensorflow_intel==2.14.0
